requirements.txt
Flask
line-bot-sdk
gunicorn
google-cloud-vision
openai==0.28
requests
beautifulsoup4
python-dotenv

